Make Update Tasks unique.
If any given Update Task (e.g. Configuration, Downloads, Application) is
on-going another one should not even be queued, let alone be run.

The above is currently achieved by means of a hack.
Me and Jakub are working on Task based resouce names.

Convert CamelCase to snake_case. This is going to be the standard naming as agreed
by everyone now.

Improve general naming style (as advised by Jakub).

Ticket: #7712

Former-commit-id: ae6ded9f3303aababcbda7a0b29445260ac27e61